chore(web): move `someday.util` to `someday.event.util` in `event/`
