+++ packages/web/src/components/ContextMenu/ContextMenuItems.tsx
@@ -5,7 +5,7 @@ import { ID_CONTEXT_MENU_ITEMS } from "@web/common/constants/web.constants";
 import { colorByPriority } from "@web/common/styles/theme.util";
 import { Schema_GridEvent } from "@web/common/types/web.event.types";
 import { assembleGridEvent } from "@web/common/utils/event/event.util";
-import { getSomedayEventCategory } from "@web/common/utils/someday.util";
+import { getSomedayEventCategory } from "@web/common/utils/event/someday.event.util";
 import {
   MenuItem,
   MenuItemLabel,
