+++ packages/web/src/ducks/events/sagas/someday.sagas.ts
@@ -3,7 +3,7 @@ import { call, put } from "redux-saga/effects";
 import { Schema_Event } from "@core/types/event.types";
 import { Schema_OptimisticEvent } from "@web/common/types/web.event.types";
 import { handleError } from "@web/common/utils/event/event.util";
-import { setSomedayEventsOrder } from "@web/common/utils/someday.util";
+import { setSomedayEventsOrder } from "@web/common/utils/event/someday.event.util";
 import { EventApi } from "@web/ducks/events/event.api";
 import {
   Action_ConvertEvent,
